BYOIP IP's no longer support update or re-create behaviors
BYOIP IPs are specified by a customer.  If they are updated, the workflow requires that we delete the accelerator and create a new one.  We do not yet support this fully as an accelerator can not have the same IP as another accelerator.  This creates some edge cases for our customers.  So instead of partially supporting, we have opted to not support to start.  We can open this door up as we better understand customer use cases and desired behaviors.