Changed TrafficDialPercentage to number json type (#26)
TrafficDialPercentage in the AGA code base is a float not an integer.  This updates the type to a float.  Also, it was realized we were not actually setting traffic dial percentage at all in the tests, so we were always getting the default value of 100, even if a customer specified a value.  We no abide by what the customer specified.