endpoint group region is now writeOnlyProperty
Endpoint Group Region can not be updated through our user interface today, but we were allowing customers to change this value.  When they changed it, we would have no side-effect although we would succeed the update.  This has been resolved so that if a customer changes this value it attempts to create an entirely new resource.  The use cases for this are limited as the endpoints inside the endpoint group will not be present in the region.  Also, a fix to UpdateEndpointGroup was made to no longer NPE if TrafficDialPercentage was null.